new code.js
While the reverse() method is useful, it doesn’t hurt to learn how to use something without the use of a built-in method. This function, in particular, is useful to help me learn how to use the for-loop to iterate through an array backward. A lot of loop examples loop from the beginning but not much from the end.